+++ jaseci_core/jaseci/api/interface.py
@@ -279,6 +279,6 @@ class Interface:
         if glob_id:
             snt = wlk._h.get_obj(wlk._m_id, uuid.UUID(glob_id))
             if snt:
-                glob_wlk = snt.walker_ids.get_obj_by_name(wlk.name)
+                glob_wlk = snt.run_architype(name=wlk.name, kind="walker")
                 if glob_wlk and glob_wlk.code_sig != wlk.code_sig:
                     wlk.apply_ir(glob_wlk.code_ir)