+++ jaseci_core/jaseci/element/element.py
@@ -46,24 +46,19 @@ class Element(Hookable):
 
     def __init__(
         self,
-        m_id,
-        h,
-        parent_id=None,
         name="basic",
         kind="generic",
         auto_save=True,
-        *args,
         **kwargs,
     ):
         self.name = name
         self.kind = kind
         self.jid = uuid.uuid4().urn
-        self.j_parent = parent_id.urn if parent_id else None  # member of
         self.j_timestamp = datetime.utcnow().isoformat()
         self.j_type = camel_to_snake(type(self).__name__)
+        Hookable.__init__(self, **kwargs)
         if self.is_master():
-            m_id = self.jid
-        Hookable.__init__(self, h, m_id, *args, **kwargs)
+            self.set_master(self.jid)
         if auto_save:
             self.save()
 
@@ -74,19 +69,6 @@ class Element(Hookable):
     @id.setter
     def id(self, obj):
         self.jid = obj.urn
-
-    @property
-    def parent_id(self) -> uuid.UUID:
-        if not self.j_parent:
-            return None
-        return uuid.UUID(self.j_parent)
-
-    @parent_id.setter
-    def parent_id(self, obj: uuid.UUID):
-        if not obj:
-            self.j_parent = None
-        else:
-            self.j_parent = obj.urn
 
     @property
     def timestamp(self):
@@ -114,8 +96,8 @@ class Element(Hookable):
         Duplicates elements by creating copy with new id
         Hook override to duplicate into mem / another store
         """
-
-        dup = type(self)(m_id=self._m_id, h=self._h, persist=persist_dup)
+        kwargs = {"m_id": self._m_id, "h": self._h, "persist": persist_dup}
+        dup = type(self)(**kwargs)
         id_save = dup.id
         for i in dup.__dict__.keys():
             if type(dup.__dict__[i]) == IdList:
@@ -167,7 +149,6 @@ class Element(Hookable):
             "jid",
             "j_type",
             "context",
-            "anchor",
             "j_timestamp",
             "version",
         ]
@@ -177,9 +158,8 @@ class Element(Hookable):
                     continue
                 jdict[i] = copy.copy(vars(self)[i])
                 if not detailed and i == "context":
-                    if "_private" in jdict[i].keys():
-                        for j in jdict[i]["_private"]:
-                            del jdict[i][j]
+                    for j in self.private_values():
+                        del jdict[i][j]
                 if deep > 0 and isinstance(jdict[i], IdList):
                     for j in range(len(jdict[i])):
                         jdict[i][j] = copy.copy(