+++ jaseci_core/jaseci/jac/interpreter/interp.py
@@ -35,57 +35,29 @@ class Interp(MachineState):
 
     def run_has_stmt(self, jac_ast, obj):
         """
-        has_stmt:
-                KW_HAS KW_PRIVATE? KW_ANCHOR? has_assign
-                (COMMA has_assign)* SEMI;
+        has_stmt: KW_HAS has_assign (COMMA has_assign)* SEMI;
         """
         kid = self.set_cur_ast(jac_ast)
-        kid = kid[1:]
-        is_private = False
-        is_anchor = False
-        while True:
-            if kid[0].name == "KW_PRIVATE":
-                kid = kid[1:]
-                is_private = True
-            if kid[0].name == "KW_ANCHOR":
-                kid = kid[1:]
-                is_anchor = True
-            self.run_has_assign(kid[0], obj, is_private, is_anchor)
-            kid = kid[1:]
-            if not len(kid) or kid[0].name != "COMMA":
-                break
-            else:
-                kid = kid[1:]
+        for i in kid:
+            if i.name == "has_assign":
+                self.run_has_assign(i, obj)
 
-    def run_has_assign(self, jac_ast, obj, is_private, is_anchor):
+    def run_has_assign(self, jac_ast, obj):
         """
-        has_assign: NAME | NAME EQ expression;
+        has_assign: KW_PRIVATE? KW_ANCHOR? (NAME | NAME EQ expression);
         """
         kid = self.set_cur_ast(jac_ast)
+        while kid[0].name in ["KW_PRIVATE", "KW_ANCHOR"]:
+            kid = kid[1:]
         var_name = kid[0].token_text()
         var_val = None  # jac's null
         if len(kid) > 1:
             var_val = self.run_expression(kid[2]).value
-        if is_anchor:
-            if "anchor" in dir(obj):
-                if obj.anchor is None:
-                    obj.anchor = var_name
-            else:
-                self.rt_error("anchors not allowed for this type", kid[0])
-
-        if var_name == "_private":
-            self.rt_error("Has variable name of `_private` not allowed!", kid[0])
         # Runs only once for walkers
-        elif var_name not in obj.context.keys() or obj.j_type != "walker":
+        if var_name not in obj.context.keys() or obj.j_type != "walker":
             JacValue(self, ctx=obj, name=var_name, value=var_val).write(
                 kid[0], force=True
             )
-        if is_private:
-            if "_private" in obj.context.keys():
-                if var_name not in obj.context["_private"]:
-                    obj.context["_private"].append(var_name)
-            else:
-                obj.context["_private"] = [var_name]
 
     def run_can_stmt(self, jac_ast, obj):
         """
@@ -863,7 +835,7 @@ class Interp(MachineState):
         ability_op: DBL_COLON | DBL_COLON NAME COLON;
         """
         kid = self.set_cur_ast(jac_ast)
-        base_arch = self.get_arch_for(atom_res.value)
+        base_arch = self.parent().get_arch_for(atom_res.value)
         if len(kid) > 1:
             kind = atom_res.value.kind
             name = kid[1].token_text()
@@ -1271,7 +1243,7 @@ class Interp(MachineState):
             result = JacSet()
             if len(kid) > 1:
                 for i in self.viable_nodes().obj_list():
-                    if self.get_arch_for(i).is_instance(kid[2].token_text()):
+                    if self.parent().get_arch_for(i).is_instance(kid[2].token_text()):
                         result.add_obj(i)
             else:
                 result += self.viable_nodes()
@@ -1289,7 +1261,7 @@ class Interp(MachineState):
         name = kid[2].token_text()
         wlk = self.yielded_walkers_ids.get_obj_by_name(name, silent=True)
         if wlk is None:
-            wlk = self.parent().spawn_walker(name, caller=self)
+            wlk = self.parent().run_architype(name=name, kind="walker", caller=self)
         if wlk is None:
             self.rt_error(f"No walker {name} exists!", kid[2])
         return wlk
@@ -1321,7 +1293,12 @@ class Interp(MachineState):
                 elif kid[0].kid[3].name == "filter_ctx":
                     self.rt_error("Filtering not allowed here", kid[0].kid[3])
             else:
-                result = Edge(m_id=self._m_id, h=self._h, kind="edge", name="generic")
+                result = Edge(
+                    m_id=self._m_id,
+                    h=self._h,
+                    kind="edge",
+                    name="generic",
+                )
             return result
 
     def run_edge_to(self, jac_ast):
@@ -1335,7 +1312,7 @@ class Interp(MachineState):
         for i in (
             self.current_node.outbound_edges() + self.current_node.bidirected_edges()
         ):
-            if len(kid) > 2 and not self.get_arch_for(i).is_instance(
+            if len(kid) > 2 and not self.parent().get_arch_for(i).is_instance(
                 kid[2].token_text()
             ):
                 continue
@@ -1357,7 +1334,7 @@ class Interp(MachineState):
         for i in (
             self.current_node.inbound_edges() + self.current_node.bidirected_edges()
         ):
-            if len(kid) > 2 and not self.get_arch_for(i).is_instance(
+            if len(kid) > 2 and not self.parent().get_arch_for(i).is_instance(
                 kid[2].token_text()
             ):
                 continue
@@ -1378,7 +1355,7 @@ class Interp(MachineState):
         kid = self.set_cur_ast(jac_ast)
         result = JacSet()
         for i in self.current_node.attached_edges():
-            if len(kid) > 2 and not self.get_arch_for(i).is_instance(
+            if len(kid) > 2 and not self.parent().get_arch_for(i).is_instance(
                 kid[2].token_text()
             ):
                 continue
@@ -1698,7 +1675,7 @@ class Interp(MachineState):
     def call_ability(self, nd, name, act_list):
         m = Interp(parent_override=self.parent(), caller=self)
         m.current_node = nd
-        arch = self.get_arch_for(nd)
+        arch = self.parent().get_arch_for(nd)
         m.push_scope(
             JacScope(parent=self, has_obj=nd, action_sets=[arch.get_all_actions()])
         )