+++ jaseci_core/jaseci/jsctl/tests/test_jsctl.py
@@ -120,8 +120,8 @@ class JsctlTest(TestCaseHelper, TestCase):
         num = len(self.call_cast("alias list"))
 
         self.call("architype delete zsb:architype:bot")
-        self.call("walker delete zsb:walker:similar_questions")
-        self.call("walker delete zsb:walker:create_nugget")
+        self.call("architype delete zsb:walker:similar_questions")
+        self.call("architype delete zsb:walker:create_nugget")
         self.assertEqual(num - 3, len(self.call_cast("alias list")))
 
     def test_jsctl_config_cmds(self):
@@ -177,15 +177,15 @@ class JsctlTest(TestCaseHelper, TestCase):
         self.call(
             "sentinel register jaseci/jsctl/tests/zsb.jac -name zsb -set_active true"
         )
-        r = self.call_cast("walker get zsb:walker:pubinit -mode keys")
+        wjid = self.call_cast("walker spawn create pubinit")["jid"]
+        r = self.call_cast(f"walker get {wjid} -mode keys")
         key = r["anyone"]
         r = self.call_cast("alias list")
-        walk = r["zsb:walker:pubinit"]
         nd = r["active:graph"]
-        r = self.call_cast(f"walker summon {walk} -key {key} -nd {nd}")
+        r = self.call_cast(f"walker summon {wjid} -key {key} -nd {nd}")
         self.assertEqual(len(r["report"]), 0)
         key = "aaaaaaaa"
-        r = self.call_cast(f"walker summon {walk} -key {key} -nd {nd}")
+        r = self.call_cast(f"walker summon {wjid} -key {key} -nd {nd}")
         self.assertFalse(r["success"])
 
     def test_jsctl_import(self):