+++ jaseci_core/jaseci/svc/jsorc/promon/__init__.py
@@ -1,1 +1,0 @@
-from .promon import Promon